Revert "feature: import .docx files"
