+++ composer.json
@@ -23,7 +23,6 @@
         "nelmio/cors-bundle": "^2.5",
         "nelmio/security-bundle": "^3.5.1",
         "phpdocumentor/reflection-docblock": "^5.6.2",
-        "phpoffice/phpword": "^1.4",
         "phpstan/phpdoc-parser": "^2.1",
         "runtime/frankenphp-symfony": "^0.2.0",
         "stof/doctrine-extensions-bundle": "^1.14",
